Added original tar.gz assembly to distribution.
The distribution module now allows one to rebuild a binary
closer to the upstream release, in tar.gz format.

To keep permissions on the backend valid, the .zip format
changed to .tar.gz.

Also, since there are no integration tests currently, the
auto-execution of the tomcat6-maven-plugin was removed
until tests are in place, and a proper disposable database
solution is viable.

--HG--
branch : maven